Clean up Signing children
